Summary in description
- Fixed the bug: if I just change the head value, then it saves in the db. No matter how many times I reload the page, I can still change the head value and it will continuously save into db the new head value.
However, if I also give est_start a value (datetime format btw) and not reload the page, then the est_start's value passes to db, and if I change head values they also pass to db. HOWEVER, once I leave the page or reload it, I cannot change the value of head field for some reason anymore; but I would still be able to change the est_start and the new value of it is continually stored in the db. Also, I noticed that if I change the est_start, then I can change the head field too. But without changing est_start first, I cannot change the head filed, instead it shows the traceback error

- Fixed the bug: If for example I choose the datetime (using built-in dropdown) to be this on the website: "04/30/2024 16:35", it saves in the db as "2024-04-30 16:35:00". BUT, on the reload of the page, it changes to "04/30/2024 09:35", but the est_start in db is still correct at "2024-04-30 16:35:00"

- Now, we are able to change the est_start on head pages (not sure if WDOs need that, but they can)